+++ src/NTMinerDataObjects/MinerServer/CoinSnapshotData.cs
@@ -8,7 +8,7 @@ namespace NTMiner.MinerServer {
         [BsonId(autoId: true)]
         public ObjectId Id { get; set; }
         public string CoinCode { get; set; }
-        public long Speed { get; set; }
+        public double Speed { get; set; }
         public int ShareDelta { get; set; }
         public int MainCoinOnlineCount { get; set; }
         public int MainCoinMiningCount { get; set; }
